requirements.txt
transformers
Pillow
fastapi
uvicorn
python-multipart